% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_data.R
\name{archive_data}
\alias{archive_data}
\title{Batch process source data}
\usage{
archive_data(limit = 100)
}
\arguments{
\item{limit}{sets a cap on the number of files to be processed at a time}
}
\value{
Returns \code{TRUE} (invisibly) upon successful completion
}
\description{
This function identifies new source data files saved in
\code{getOption('zeitgebeR')$hue_storage_path} and
\code{getOption('zeitgebeR')$darksky_storage_path}. Each file is loaded and
parsed, and the data are appended to an RDS file. It then moves the processed
source files into a zip archive. Upon completion, the storage paths should
contain only an RDS file and a zip archive.
}
