% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_from_models}
\alias{predict_from_models}
\title{Predict the ideal state}
\usage{
predict_from_models(models, state, weather)
}
\arguments{
\item{models}{the list of models to predict from, like those returned by
\code{\link{fit_models}}}

\item{state}{the current Hue state, as returned by
\code{\link{parse_hue_state}}}

\item{weather}{current weather, as returned by \code{\link{parse_weather}}}
}
\value{
Returns a \code{\link[dplyr]{data_frame}} of light information
  including \code{bri} and \code{ct} values predicted from the supplied
  models.
}
\description{
Predict the ideal state
}
