% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{scale}
\alias{scale}
\alias{scale_bri}
\alias{unscale_bri}
\alias{scale_ct}
\alias{unscale_ct}
\title{Scale and unscale brightness and color temperature}
\usage{
scale_bri(bri)

unscale_bri(bri)

scale_ct(ct)

unscale_ct(ct)
}
\arguments{
\item{bri}{the brightness value to scale or unscale}

\item{ct}{the color temperature value to scale or unscale}
}
\value{
Returns scaled (i.e. [0, 1]) values or unscaled (brightness in
  [1, 254]; color temperature in [153, 500]) values.
}
\description{
The quasibinomial models created by \code{\link{fit_models}} produce
estimates in the range [0, 1], but brightness and color temperature are in
the ranges [1, 254] and [153, 500], respectively. The following functions are
provided to help convert to and from the model scale.
}
