% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_configuration}
\alias{set_configuration}
\alias{reset_configuration}
\title{Configure Environment}
\usage{
set_configuration(hue_bridge_ip, hue_username, darksky_key, lat, lon,
  hue_storage_path, darksky_storage_path)

reset_configuration()
}
\arguments{
\item{hue_bridge_ip}{the IP address of your Hue Bridge}

\item{hue_username}{the username with access to your Hue Bridge}

\item{darksky_key}{Dark Sky secret key}

\item{lat}{latitude}

\item{lon}{longitude}

\item{hue_storage_path}{path to directory where Hue training data are stored}

\item{darksky_storage_path}{path to directory where Dark Sky weather data are
stored}
}
\value{
Returns \code{TRUE} (invisibly) if configuration was set succesfully.
}
\description{
These functions are used primarily for their side effects -- specifically to
set (or reset) secrets and other configuration details.
}
