% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{source_data}
\alias{source_data}
\alias{record_hue_state}
\alias{record_weather}
\alias{parse_hue_state}
\alias{parse_weather}
\title{Functions to record and parse source data}
\usage{
record_hue_state(cache_timeout = 0)

record_weather(cache_timeout = 15)

parse_hue_state(x)

parse_weather(x)
}
\arguments{
\item{cache_timeout}{time (in minutes) to reuse cached results before hitting
the API again}

\item{x}{the object to parse, i.e.
\code{x = yaml::yaml.load_file(path)}) where \code{path} is a data file
saved by \code{\link[=record_hue_state]{record_hue_state()}} or \code{\link[=record_weather]{record_weather()}}}
}
\value{
\code{\link[=record_hue_state]{record_hue_state()}} and \code{\link[=record_weather]{record_weather()}}
return \code{TRUE} (invisibly) upon success. \code{\link[=parse_hue_state]{parse_hue_state()}}
and \code{\link[=parse_weather]{parse_weather()}} return a \code{\link[dplyr:data_frame]{dplyr::data_frame()}}
of relevant features.
}
\description{
Functions to record and parse source data
}
