% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_data.R
\name{training_data}
\alias{training_data}
\alias{record_hue_state}
\alias{record_weather}
\alias{record_data}
\title{Functions to record and manipulate training data}
\usage{
record_hue_state()

record_weather(cache_timeout = 15)

record_data(cache_timeout = 15)
}
\arguments{
\item{cache_timeout}{time (in minutes) to cache results before hitting the
API again}
}
\value{
Returns `invisible(TRUE)` upon success.
}
\description{
Functions to record and manipulate training data
}
