% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitgebeR-package.R
\docType{package}
\name{zeitgebeR-package}
\alias{zeitgebeR}
\alias{zeitgebeR-package}
\title{zeitgebeR: Machine Learning Enhancements for the Philips Hue Lighting System}
\description{
This package contains functions to support a machine learning
pipeline for your Philips Hue lighting system that continuously adjusts the
lighting to match the time of day, current weather conditions, and learned
preferences. See the README at
<https://github.com/fascinatingfingers/zeitgebeR/README.md>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/fascinatingfingers/zeitgeber}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/zeitgeber/issues}
}

}
\keyword{internal}
