% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitgeber.R
\encoding{UTF-8}
\name{zeitgeber}
\alias{zeitgeber}
\title{Create Table of Time Cues}
\usage{
zeitgeber(time = Sys.time(), lat = getOption("DarkSky")$lat,
  lon = getOption("DarkSky")$lon)
}
\arguments{
\item{time}{vector of date/time values}

\item{lat}{latitude (in decimal degrees). Positive north; negative south.}

\item{lon}{longitude (in decimal degrees). Positive east; negative west.}
}
\value{
Returns a data_frame with the following diurnal features:
  \itemize{
    \item \code{datetime}
    \item \code{solar_elevation} (in degrees)
    \item \code{solar_azimuth} (in degrees)
    \item \code{solar_azimuth_x} east--west movement, scaled [-1, 1]
    \item \code{solar_azimuth_y} north--south movement, scaled [-1, 1]
    \item \code{civil_daylight} \code{TRUE} when the solar elevation is
      greater than -6°
    \item \code{nautical_daylight} \code{TRUE} when the solar elevation is
      greater than -12°
    \item \code{astronomical_daylight} \code{TRUE} when the solar elevation
      is greater than -18°
    \item \code{working_hours} \code{TRUE} from 9a--5p; 10a--6p on weekends
    \item \code{awake_hours} \code{TRUE} from 8a--12a; 9a--1a on weekends
  }
}
\description{
Zeitgeber, Geman for \emph{time-giver}, are external cues that help
synchronize an organism's biolgical rhythms to Earth's day/night cycles and
seasons. This eponymous function takes a vector of date/time values (plus
latitude/longitude values retrieved from the session options; see Details)
and returns a \code{\link[dplyr]{data_frame}} of zeitgeber features.
}
\details{
This function retrieves location information saved in
\code{options()$DarkSky}. Before calling this function, set your credentials
and other configuration details with a call to \code{set_configuration}.
}
