% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit Brightness and Color Temperature Models}
\usage{
fit_models(dta = training_data())
}
\arguments{
\item{dta}{training dataset, like that returned by
\code{\link{training_data}}}
}
\value{
Returns a named list with two models, one for brightness and one for
  color temperature.
}
\description{
After recording data and preparing a dataset with \code{\link{training_data}},
the \code{fit_models} function will fit two generalized linear models: one
to predict brightness for the given features, and the other to predict color
temperature.
}
