% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{set_configuration}
\alias{set_configuration}
\alias{reset_configuration}
\title{Configure Environment}
\usage{
set_configuration(hue_bridge_ip, hue_username, darksky_key, lat, lon,
  hue_storage_path = "data/state", darksky_storage_path = "data/weather")

reset_configuration()
}
\arguments{
\item{hue_bridge_ip}{the IP address of your Hue Bridge}

\item{hue_username}{the username with access to your Hue Bridge}

\item{darksky_key}{Dark Sky secret key}

\item{lat}{latitude (in decimal degrees). Positive north; negative south.}

\item{lon}{longitude (in decimal degrees). Positive east; negative west.}

\item{hue_storage_path}{path to directory where Hue training data are stored}

\item{darksky_storage_path}{path to directory where Dark Sky weather data are
stored}
}
\value{
Returns \code{TRUE} (invisibly) if options were successfully set or
reset.
}
\description{
These functions are used specifically for their side effects -- namely to
set (or reset) authentication secrets and configuration details stored in
\code{options()}.
}
\seealso{
\code{\link[PhilipsHue:set_bridge_credentials]{PhilipsHue::set_bridge_credentials()}}

\code{\link[DarkSky:set_dark_sky_credentials]{DarkSky::set_dark_sky_credentials()}}
}
