% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{training_data}
\alias{training_data}
\title{Prepare training data}
\usage{
training_data(state, weather)
}
\arguments{
\item{state}{\code{\link[dplyr]{data_frame}} of Hue state observations,
typically created by \code{\link{archive_data}}}

\item{weather}{\code{\link[dplyr]{data_frame}} of weather observations,
typically created by \code{\link{archive_data}}}
}
\value{
Returns a \code{\link[dplyr]{data_frame}} ready for a call to
  \code{\link{fit_models}}
}
\description{
This function combines historical observations of Hue state and weather
conditions into a \code{\link[dplyr]{data_frame}} formatted for model
fitting.
}
