% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitgeber.R
\name{zeitgeber}
\alias{zeitgeber}
\title{Create Table of Time Cues}
\usage{
zeitgeber(time = Sys.time(), lat = getOption("DarkSky")$lat,
  lon = getOption("DarkSky")$lon)
}
\arguments{
\item{time}{time}

\item{lat}{latitude}

\item{lon}{longitude}
}
\value{
Returns a dataframe with `length(x)` rows and columns of diurnal
  features
}
\description{
Create Table of Time Cues
}
